@@ -18,19 +18,19 @@ environment:
 
 dependencies:
   # mailer2: ^1.2.5
-  mailer: ^3.2.1
+  mailer: ^5.0.0
 
-  flutter_email_sender: ^3.0.1
-  webview_flutter: ^0.3.22+1
+  flutter_email_sender: ^5.0.0
+  webview_flutter: ^2.0.4
   flutter_webview_plugin: ^0.3.0+2
-  http: ^0.12.1
+  http: ^0.13.1
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  shared_preferences: ^0.5.7+3
+  cupertino_icons: ^1.0.2
+  shared_preferences: ^2.0.5
 
 dev_dependencies:
   flutter_test:
